Update FirewallTests-FiskalyCloud2.csv
Previously it wasn't possible to access "kassensichv-middleware.fiskaly.com/api/v2" and the way to check if the site is reachable was to wait for Code 401. But now the site is accessible  and the response Code should be 200. The Code 401 isn't returned anymore, so the Script always fails for Fiskaly V2. With Code 200 it works like it was supposed to.