Corrected moment of SwissbitCloud-TSE (FCC) initialization
The FCC is already initialized with the first start of the middleware and not just with the initial-operation-receipt. With the old version of this document it seemed like the SwissbitCloud-TSE can be installed on other devices as long as no initial-operation-receipt was sent. The english version seems correct in that matter.